refactor code, introduce BCs, license is now MIT
refs #13, refs #14
