+++ LICENSE.md
@@ -1,18 +1,22 @@
 # License
 
-<a rel="license"
-href="http://creativecommons.org/licenses/by-sa/3.0/deed.en_US"><img
-alt="Creative Commons License" style="border-width:0"
-src="http://i.creativecommons.org/l/by-sa/3.0/88x31.png" /></a><br /><span
-xmlns:dct="http://purl.org/dc/terms/" property="dct:title">params</span>
-is licensed under a
-<a rel="license" href="http://creativecommons.org/licenses/by-sa/3.0/deed.en_US">
-Creative Commons Attribution-ShareAlike 3.0 Unported License
-</a>.
+Copyright (c) 2014-2015 Steffen Gransow
 
-C-BY-SA-3.0 means, you are free to share, remix and make commercial use of the
-work as long as you attribute and share alike. See linked license for details.
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is furnished
+to do so, subject to the following conditions:
 
-The authors and contributors are mentioned in [github contributors graph](graphs/contributors)
-of this repository. The license file must be left intact for attribution and
-sharing of this work.
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+THE SOFTWARE.
+
