+++ composer.json
@@ -1,10 +1,10 @@
 {
     "name": "graste/params",
     "type": "library",
-    "description": "Array wrapper that eases the retrieval of values. Has parameters, options and settings traits for inclusion in other libs.",
+    "description": "Array wrapper that eases the retrieval of values. Has parameters, options and settings and traits for inclusion in other libraries.",
     "keywords": ["parameters", "options", "settings", "array", "search", "library"],
     "homepage": "https://github.com/graste/params",
-    "license": "CC-BY-SA-3.0",
+    "license": "MIT",
     "authors": [
         {
             "name": "Steffen Gransow",
@@ -22,7 +22,7 @@
     "prefer-stable": true,
     "require": {
         "php": ">=5.4.0",
-        "mtdowling/jmespath.php": ">=1.1.0 <=2.1.0"
+        "mtdowling/jmespath.php": ">=1.1.0"
     },
     "require-dev": {
         "phpunit/phpunit": "~4.3.0",
@@ -32,13 +32,13 @@
         "phine/sami-github": "~1.0.4"
     },
     "autoload": {
-        "psr-0": {
+        "psr-4": {
             "Params\\": "src/"
         }
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "2.1-dev"
+            "dev-master": "3.0-dev"
         }
     }
 }
