+++ src/Params/OptionsMutableInterface.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Params;
+
+/**
+ * Defines methods for mutable options.
+ */
+interface OptionsMutableInterface extends ParamsMutableInterface
+{
+
+}
