+++ src/Params/ParametersMutableInterface.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Params;
+
+/**
+ * Defines methods for mutable parameters.
+ */
+interface ParametersMutableInterface extends ParamsMutableInterface
+{
+
+}
