+++ src/Params/SettingsMutableInterface.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Params;
+
+/**
+ * Defines methods for mutable settings.
+ */
+interface SettingsMutableInterface extends ParamsMutableInterface
+{
+
+}
